requirements.txt
gradio
google-generativeai
transformers
torch
scipy